[REFACTOR] Cli file structure
