+++ cli/src/commands/utils.rs
