+++ cli/src/main.rs
@@ -1,14 +1,15 @@
-mod cli;
-mod models;
-mod version;
-mod utils;
-mod config;
-
 use clap::Parser;
-use cli::errors::CliError;
 use nanocl_client::NanoclClient;
 
-use models::*;
+mod utils;
+mod error;
+mod config;
+mod models;
+mod version;
+mod commands;
+
+use error::CliError;
+use models::{Cli, Commands};
 
 fn process_error(args: &Cli, err: CliError) {
   match err {
@@ -35,10 +36,10 @@ fn process_error(args: &Cli, err: CliError) {
 async fn execute_args(args: &Cli) -> Result<(), CliError> {
   let client = NanoclClient::connect_with_unix_default().await;
   match &args.command {
-    Commands::Setup(args) => cli::exec_setup(args).await,
-    Commands::Namespace(args) => cli::exec_namespace(&client, args).await,
-    Commands::Cargo(args) => cli::exec_cargo(&client, args).await,
-    Commands::Version => cli::exec_version(&client).await,
+    Commands::Setup(args) => commands::exec_setup(args).await,
+    Commands::Namespace(args) => commands::exec_namespace(&client, args).await,
+    Commands::Cargo(args) => commands::exec_cargo(&client, args).await,
+    Commands::Version => commands::exec_version(&client).await,
   }
 }
